# backend/swagger.yaml
swagger: "2.0"
info:
  description: "Trainer REST API documentation"
  version: "1.0.0"
  title: "Trainer"
host: "trainer.ksi.fit.cvut.cz"
basePath: "/api"
...
schemes:
- "https"
paths:
...
  /lessons/{id}/modules/{moduleId}:
    put:
      tags:
        - "lesson-modules"
      summary: "Add module to lesson"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "path"
          description: "Identifier of module to add / modify"
          name: "moduleId"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "Description of LessonModule properties"
          required: true
          schema:
            $ref: "#/definitions/LessonModuleEditDTO"
      responses:
        "200":
          description: "Success, created/updated LessonModule"
          schema:
            $ref: "#/definitions/LessonModuleReadDTO"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to access lesson or module"
        "404":
          description: "Lesson or module with given identifier not found"
      security:
        - login_secret: [ ]
    delete:
      tags:
        - "lesson-modules"
      summary: "Remove module from lesson"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          description: "Unique lesson identifier"
          name: "id"
          type: "integer"
          required: true
        - in: "path"
          description: "Identifier of module to remove"
          name: "moduleId"
          type: "integer"
          required: true
      responses:
        "204":
          description: "Success, removed LessonModule"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to access lesson or module"
        "404":
          description: "Lesson or module with given identifier not found"
      security:
        - login_secret: [ ]

  # Modules
...
  /lessons/{id}/modules/{moduleId}/data/me:
    put:
      tags:
        - "student-modules"
      summary: "Replaces existing StudentModule"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Unique identifier of lesson"
          type: "integer"
          required: true
        - in: "path"
          name: "moduleId"
          description: "Unique identifier of module"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "StudentModule to be created / values to be updated"
          required: true
          schema:
            $ref: "#/definitions/StudentModuleEditDTO"
      responses:
        "200":
          description: "Success, StudentModule created / edited"
          schema:
            $ref: "#/definitions/StudentModuleReadDTO"
        "401":
          description: "Unauthorized user"
        "404":
          description: "Lesson or module with given identifier not found"
      security:
        - login_secret: [ ]
    delete:
      tags:
        - "student-modules"
      summary: "Deletes student module data"
      parameters:
        - in: "path"
          name: "id"
          description: "Unique identifier of lesson"
          type: "integer"
          required: true
        - in: "path"
          name: "moduleId"
          description: "Unique identifier of module"
          type: "integer"
          required: true
      responses:
        "204":
          description: "Success, student module data deleted"
        "401":
          description: "Unauthorized user"
        "403":
          description: "Insufficient rights to delete student module data"
        "404":
          description: "Lesson or module or student module not found"
      security:
        - login_secret: [ ]
  /lessons/{id}/modules/{moduleId}/file/me:
    post:
      tags:
        - "student-modules"
      summary: "Upload student module file"
      consumes:
        - "multipart/form-data"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Unique identifier of lesson"
          type: "integer"
          required: true
        - in: "path"
          name: "moduleId"
          description: "Unique identifier of module"
          type: "integer"
          required: true
        - in: "formData"
          name: "file"
          type: "file"
          format: "image/*"
          description: "File to be uploaded"
          required: true
      responses:
        "200":
          description: "Success, student module file"
          schema:
            $ref: "#/definitions/StudentModuleReadDTO"
        "401":
          description: "Unauthorized user"
        "404":
          description: "Lesson or module with given identifier not found"
      security:
        - login_secret: [ ]
    get:
      tags:
        - "student-modules"
      summary: "Get own student module file"
      produces:
        - "application/x-tar"
      parameters:
        - in: "path"
          name: "id"
          description: "Unique identifier of lesson"
          type: "integer"
          required: true
        - in: "path"
          name: "moduleId"
          description: "Unique identifier of module"
          type: "integer"
          required: true
      responses:
        "200":
          description: "Success, own student module file"
        "401":
          description: "Unauthorized user"
        "404":
          description: "Lesson or module with given identifier not found or file does not exist"
      security:
        - login_secret: [ ]
  /lessons/{id}/modules/{moduleId}/file/users/{userId}:
    get:
      tags:
        - "student-modules"
      summary: "Get given student module file"
      produces:
        - "application/x-tar"
      parameters:
        - in: "path"
          name: "id"
          description: "Unique identifier of lesson"
          type: "integer"
          required: true
        - in: "path"
          name: "moduleId"
          description: "Unique identifier of module"
          type: "integer"
          required: true
        - in: "path"
          name: "userId"
          description: "Unique identifier of user"
          type: "integer"
          required: true
      responses:
        "200":
          description: "Success, student's module file"
        "401":
          description: "Unauthorized user"
        "404":
          description: "Lesson or module or user with given identifier not found or file does not exist"
      security:
        - login_secret: [ ]
  /lessons/{id}/modules/{moduleId}/requests/me:
    put:
      tags:
        - "student-modules"
      summary: "Replaces existing StudentModuleRequest"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Unique identifier of lesson"
          type: "integer"
          required: true
        - in: "path"
          name: "moduleId"
          description: "Unique identifier of module"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "StudentModuleRequest to be created / values to be updated"
          required: true
          schema:
            $ref: "#/definitions/StudentModuleRequestEditDTO"
      responses:
        "204":
          description: "Success, StudentModuleRequest created / edited"
        "401":
          description: "Unauthorized user"
        "404":
          description: "Lesson or module with given identifier not found"
      security:
        - login_secret: [ ]
    delete:
      tags:
        - "student-modules"
      summary: "Deletes student module request"
      parameters:
        - in: "path"
          name: "id"
          description: "Unique identifier of lesson"
          type: "integer"
          required: true
        - in: "path"
          name: "moduleId"
          description: "Unique identifier of module"
          type: "integer"
          required: true
      responses:
        "204":
          description: "Success, student module request deleted"
        "401":
          description: "Unauthorized user"
        "404":
          description: "Lesson or module or student module request not found"
      security:
        - login_secret: [ ]
  /lessons/{id}/modules/{moduleId}/requests/users/{userId}:
    put:
      tags:
        - "student-modules"
      summary: "Replaces existing StudentModuleRequest (teacher side)"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "Unique identifier of lesson"
          type: "integer"
          required: true
        - in: "path"
          name: "moduleId"
          description: "Unique identifier of module"
          type: "integer"
          required: true
        - in: "path"
          name: "userId"
          description: "Unique identifier of student"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "StudentModuleRequest to be created / values to be updated"
          required: true
          schema:
            $ref: "#/definitions/StudentModuleRequestTeacherDTO"
      responses:
        "200":
          description: "Success, StudentModuleRequest created / edited"
          schema:
            $ref: "#/definitions/StudentModuleReadDTO"
        "401":
          description: "Unauthorized user"
        "404":
          description: "Lesson or module with given identifier not found"
      security:
        - login_secret: [ ]

  # Notifications
...
securityDefinitions:
  login_secret:
    type: "apiKey"
    name: "LoginSecret"
    in: "header"
definitions:
...
  LessonModuleReadDTO:
    type: "object"
    description: "LessonModule description"
    properties:
      id:
        type: "integer"
        description: "Unique identifier of LessonModule"
        example: 44
      name:
        type: "string"
        description: "Name of module"
        example: "Compilation – task"
      type:
        type: "string"
        enum: [ "TEXT", "QUIZ", "CODE", "ASSIGNMENT" ]
        description: "Type of module"
        example: "ASSIGNMENT"
      order:
        type: "integer"
        description: "Module order in lesson"
        example: 1
  LessonModuleEditDTO:
    type: "object"
    description: "Description of module to edit / create"
    properties:
      order:
        type: "integer"
        description: "Module order in lesson"
        example: 1
      depends:
        type: "integer"
        description: "Identifier of module on which the module should depend (null if none)"
        example: 2
...
  StudentModuleRequestType:
    type: "string"
    description: "Student request type"
    enum: ["HELP", "COMMENT", "EVALUATE"]
    example: "HELP"
  StudentModuleEditDTO:
    type: "object"
    description: "Student module settings"
    properties:
      allowedShow:
        type: "boolean"
        description: "Anonymized publication of solution"
        example: false
      percent:
        type: "integer"
        description: "Percent student has reached in the module"
        example: 75
  StudentModuleReadDTO:
    type: "object"
    description: "Student module data transfer object"
    properties:
      percent:
        type: "integer"
        description: "Percent student has reached in the module"
        example: 80
      completedEarly:
        type: "boolean"
        description: "Flag indicating whether student completed the module before end of lesson"
        example: false
      completed:
        type: "string"
        format: "date-time"
        description: "Time when user completed the module (or null if uncompleted)"
        example: null
  StudentModuleRequestEditDTO:
    type: "object"
    description: "Student module request edit object"
    properties:
      requestText:
        type: "string"
        description: "Request text"
        example: "Please, help me"
      requestType:
        $ref: "#/definitions/StudentModuleRequestType"
  StudentModuleRequestTeacherDTO:
    type: "object"
    description: "Student module request teacher DTO"
    properties:
      responseText:
        type: "string"
        description: "Response to the request or comment"
        example: "You should fix line 20 on your code"
      evaluation:
        type: "boolean"
        description: "Should module get evaluated? (manualEval modules)"
        example: false
      percent:
        type: "integer"
        description: "If module should be evaluated, sets the percent to given value (manualEval modules)"
        example: null

  # Notification
...